Add prompts when installing stuff

+++ install/macvim.sh
@@ -3,7 +3,10 @@ is_osx || return 1
 
 # Install MacVim
 if ! formula_exists 'macvim'; then
-    brew install macvim --with-lua
+    prompt "Install macvim."
+    if is_confirmed; then
+        brew install macvim --with-lua
+    fi
 else
     log_arrow "MacVim is installed"
 fi
